Use Svelte for tags page
